Optimization completed
I ran the optimization file for nfe = 100.000. I saved the results to a csv file. I think the next steps from here is that we do some data analysis in a notebook and define certain critical thresholds for our actor (e.g. total costs may never exceed a certain value). Out of these threshold we can filter a few candidate solutions which we can further use to run expirements on.